Extracted error boundary component to its own file
