+++ apps/stats/src/App.tsx
@@ -1,29 +1,12 @@
 import GlobalDataProvider from './providers/GlobalDataProvider';
 import React from 'react';
+import StatsErrorBoundary from './components/errors/StatsErrorBoundary';
 import StatsErrorPage from './components/errors/StatsErrorPage';
 import {APP_ROUTE_PREFIX, routes} from '@src/routes';
 import {AppProvider, BaseAppProps, FrameworkProvider, Outlet, RouterProvider} from '@tryghost/admin-x-framework';
 import {ShadeApp} from '@tryghost/shade';
 
 export {useAppContext} from '@tryghost/admin-x-framework';
-
-class StatsErrorBoundary extends React.Component<{children: React.ReactNode}, {hasError: boolean, error?: Error}> {
-    constructor(props: {children: React.ReactNode}) {
-        super(props);
-        this.state = {hasError: false};
-    }
-
-    static getDerivedStateFromError(error: Error) {
-        return {hasError: true, error};
-    }
-
-    render() {
-        if (this.state.hasError) {
-            return <StatsErrorPage error={this.state.error} />;
-        }
-        return this.props.children;
-    }
-}
 
 const App: React.FC<BaseAppProps> = ({framework, designSystem, appSettings}) => {
     return (
